Get rid of abstract class and high and low objects
