--- src/encoded/schemas/changelogs/high_throughput_single_cell_experiment.md
@@ -1,1 +1,0 @@
-## Changelog for high_throughput_single_cell_experiment.json
