--- src/encoded/schemas/changelogs/low_throughput_single_cell_experiment.md
@@ -1,1 +1,0 @@
-## Changelog for low_throughput_single_cell_experiment.json
