+++ src/encoded/schemas/changelogs/replicate.md
@@ -7,7 +7,7 @@
 * *biological_replicate_number* and *technical_replicate_number* were set to have a minimum of 1.
 * *rbns_protein_concentration* was set to have a minimum of 0.
 * Calculated property *libraries* is removed.
-* *experiment* property list now includes HighThroughputSingleCellExperiment.json and SingleCellUnit.json
+* *experiment* property list now includes SingleCellUnit.json
 
 ### Schema version 9
