--- src/encoded/schemas/changelogs/single_cell_experiment.md
@@ -1,1 +1,0 @@
-## Changelog for single_cell_experiment.json
