--- src/encoded/schemas/high_throughput_single_cell_experiment.json
@@ -1,210 +1,0 @@
-{
-    "title": "High Throughput Single Cell Experiment",
-    "description": "Schema for submitting metadata for a high throughput single cell assay (such as 10X).",
-    "comment": "A high throughput single cell experiment is a special case of dataset. It includes assay metadata, replicate information and data files.",
-    "id": "/profiles/high_throughput_single_cell_experiment.json",
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "type": "object",
-    "required": ["award", "lab", "assay_term_name", "biosample_ontology"],
-    "identifyingProperties": ["uuid", "accession" , "aliases"],
-    "additionalProperties": false,
-    "mixinProperties": [
-        { "$ref": "mixins.json#/schema_version" },
-        { "$ref": "mixins.json#/uuid" },
-        { "$ref": "mixins.json#/accession" },
-        { "$ref": "mixins.json#/aliases" },
-        { "$ref": "mixins.json#/attribution" },
-        { "$ref": "mixins.json#/submitted" },
-        { "$ref": "mixins.json#/notes" },
-        { "$ref": "dataset.json#/properties" },
-        { "$ref": "single_cell_experiment.json#/properties" },
-        { "$ref": "mixins.json#/references" },
-        { "$ref": "mixins.json#/documents" },
-        { "$ref": "mixins.json#/biosample_ontology" },
-        { "$ref": "mixins.json#/submitter_comment" }
-    ],
-    "dependencies": {
-        "status": {
-            "oneOf": [
-                {
-                    "required": ["date_released"],
-                    "properties": {
-                        "status": {
-                            "enum": ["released", "revoked"]
-                        }
-                    }
-                },
-                {
-                    "required": ["date_submitted"],
-                    "properties": {
-                        "status": {
-                            "enum": ["submitted"]
-                        }
-                    }
-                },
-                {
-                    "not": {
-                        "properties": {
-                            "status": {
-                                "enum": ["released", "revoked", "submitted"]
-                            }
-                        }
-                    }
-                }
-            ]
-        }
-    },
-    "properties": {
-        "schema_version": {
-            "default": "1"
-        },
-        "supersedes": {
-            "title": "Supersedes" ,
-            "description": "The high throughput single cell experiment(s) that this high throughput single cell experiment supersedes by virtue of being newer, better etc. than the one(s) it supersedes.",
-            "type": "array",
-            "default": [],
-            "uniqueItems": true,
-            "items": {
-                "comment": "See high_throughput_single_cell_experiment.json for a list of available identifiers.",
-                "type": "string",
-                "linkTo": "HighThroughputSingleCellExperiment"
-            }
-        }
-    },
-    "facets": {
-        "assay_title": {
-            "title": "Assay title",
-            "type": "typeahead",
-            "open_on_load": true
-        },
-        "status": {
-            "title": "Status",
-            "open_on_load": true
-        },
-        "target.investigated_as": {
-            "title": "Target category"
-        },
-        "target.label": {
-            "title": "Target of assay",
-            "type": "typeahead",
-            "length": "long",
-            "open_on_load": true
-        },
-        "biosample_ontology.classification": {
-            "title": "Biosample classification"
-        },
-        "biosample_ontology.term_name" : {
-            "title": "Biosample term name",
-            "type": "typeahead",
-            "length": "long",
-            "open_on_load": true
-        },
-        "biosample_ontology.organ_slims": {
-            "title": "Organ",
-            "type": "typeahead"
-        },
-        "biosample_ontology.cell_slims": {
-            "title": "Cell",
-            "type": "typeahead"
-        },
-        "award.project": {
-            "title": "Project"
-        },
-        "award.rfa": {
-            "title": "RFA"
-        },
-        "assembly": {
-            "title": "Genome assembly"
-        },
-        "date_released": {
-            "title": "Date released"
-        },
-        "date_submitted": {
-            "title": "Date submitted"
-        },
-        "lab.title": {
-            "title": "Lab"
-        },
-        "internal_status": {
-            "title": "Internal status"
-        }
-    },
-    "columns": {
-        "accession": {
-            "title": "Accession"
-        },
-        "assay_term_name": {
-            "title": "Assay name"
-        },
-        "assay_title": {
-            "title": "Assay title"
-        },
-        "target.label": {
-            "title": "Target of assay"
-        },
-        "target.genes.symbol": {
-            "title": "Target gene symbol"
-        },
-        "biosample_summary": {
-            "title": "Biosample summary"
-        },
-        "biosample_ontology.term_name": {
-            "title": "Biosample term name"
-        },
-        "description": {
-            "title": "Description"
-        },
-        "lab.title": {
-            "title": "Lab"
-        },
-        "award.project": {
-            "title": "Project"
-        },
-        "status": {
-            "title": "Status"
-        }
-    },
-    "boost_values": {
-        "@type": 1.0,
-        "accession": 1.0,
-        "alternate_accessions": 1.0,
-        "assay_term_name": 1.0,
-        "assay_term_id": 1.0,
-        "dbxrefs": 1.0,
-        "aliases": 1.0,
-        "award.title": 1.0,
-        "biosample_ontology.term_id": 1.0,
-        "biosample_ontology.term_name": 1.0,
-        "biosample_ontology.classification": 1.0,
-        "biosample_ontology.organ_slims": 1.0,
-        "biosample_ontology.cell_slims": 1.0,
-        "biosample_ontology.developmental_slims": 1.0,
-        "category_slims": 1.0,
-        "objective_slims": 1.0,
-        "type_slims": 1.0,
-        "assay_synonyms": 1.0,
-        "biosample_ontology.synonyms": 1.0,
-        "award.project": 1.0,
-        "award.name": 1.0,
-        "submitted_by.email": 1.0,
-        "submitted_by.first_name": 1.0,
-        "submitted_by.last_name": 1.0,
-        "lab.institute_name": 1.0,
-        "lab.institute_label": 1.0,
-        "lab.title": 1.0,
-        "possible_controls.accession": 1.0,
-        "possible_controls.alternate_accessions": 1.0,
-        "target.aliases": 1.0,
-        "target.genes.symbol": 1.0,
-        "target.label": 1.0,
-        "target.genes.dbxrefs": 1.0,
-        "target.organism.name": 1.0,
-        "target.organism.scientific_name": 1.0,
-        "references.title": 1.0,
-        "notes": 1.0,
-        "internal_tags": 1.0,
-        "doi": 1.0,
-        "description": 8.0
-    },
-    "changelog": "/profiles/changelogs/high_throughput_single_cell_experiment.md"
-}
