+++ src/encoded/schemas/replicate.json
@@ -46,7 +46,7 @@
             "description": "The experiment the replicate belongs to.",
             "comment": "See experiment.json for available identifiers.",
             "type": "string",
-            "linkTo": ["Experiment", "FunctionalCharacterizationExperiment", "HighThroughputSingleCellExperiment", "SingleCellUnit"]
+            "linkTo": ["Experiment", "FunctionalCharacterizationExperiment", "SingleCellUnit"]
         },
         "library": {
             "title": "Library",
