--- src/encoded/schemas/single_cell_experiment.json
@@ -1,214 +1,0 @@
-{
-    "title": "Single Cell Experiment",
-    "description": "Schema for submitting metadata for a single cell assay (such as Fluidigm C1).",
-    "comment": "A single cell experiment is a special case of dataset. It includes assay metadata, replicate information and data files.",
-    "id": "/profiles/single_cell_experiment.json",
-    "$schema": "http://json-schema.org/draft-04/schema#",
-    "type": "object",
-    "required": ["award", "lab", "assay_term_name", "biosample_ontology"],
-    "identifyingProperties": ["uuid", "accession" , "aliases"],
-    "additionalProperties": false,
-    "mixinProperties": [
-        { "$ref": "mixins.json#/schema_version" },
-        { "$ref": "mixins.json#/uuid" },
-        { "$ref": "mixins.json#/accession" },
-        { "$ref": "mixins.json#/aliases" },
-        { "$ref": "mixins.json#/attribution" },
-        { "$ref": "mixins.json#/submitted" },
-        { "$ref": "mixins.json#/notes" },
-        { "$ref": "dataset.json#/properties" },
-        { "$ref": "mixins.json#/references" },
-        { "$ref": "mixins.json#/documents" },
-        { "$ref": "mixins.json#/biosample_ontology" },
-        { "$ref": "mixins.json#/submitter_comment" }
-    ],
-    "properties": {
-        "schema_version": {
-            "hidden comment": "Bump the default in the subclasses."
-        },
-        "assay_term_name": {
-            "title": "Assay name",
-            "description": "OBI (Ontology for Biomedical Investigations) ontology term for the assay.",
-            "type": "string",
-            "enum": [
-                "single-cell RNA sequencing assay",
-                "single-nucleus ATAC-seq"
-            ]
-        },
-        "date_submitted": {
-            "title": "Date submitted" ,
-            "description": "The date that the experiment received all required replicates and files",
-            "comment": "Do not submit. date_submitted will be assigned only by DCC personnel.",
-            "type": "string",
-            "format": "date",
-            "permission": "import_items"
-        },
-        "target": {
-            "title": "Target",
-            "description": "For assays, such as CRISPR screen, the name of the gene whose expression or product is under investigation for the functional characterization experiment. This property is not submittable, and examined_loci should be specified instead.",
-            "comment": "See target.json for available identifiers.",
-            "type": "string",
-            "linkTo": "Target",
-            "permission": "import_items"
-        },
-        "possible_controls": {
-            "title": "Controls",
-            "description": "Dataset that contain files that can serve as scientific controls for this functional characterization experiment.",
-            "type": "array",
-            "default": [],
-            "uniqueItems": true,
-            "items": {
-                "title": "Control",
-                "description": "A dataset that contains files that can serve as a scientific control for this functional characterization experiment.",
-                "comment": "See dataset.json for a list of available identifiers.",
-                "type": "string",
-                "linkTo": "Dataset"
-            }
-        },
-        "control_type": {
-            "type": "string",
-            "title": "Control type",
-            "description": "The specific type of control if the experiment is a control.",
-            "enum": [
-                "control"
-            ]
-        }
-    },
-    "facets": {
-        "assay_title": {
-            "title": "Assay title",
-            "type": "typeahead",
-            "open_on_load": true
-        },
-        "status": {
-            "title": "Status",
-            "open_on_load": true
-        },
-        "target.investigated_as": {
-            "title": "Target category"
-        },
-        "target.label": {
-            "title": "Target of assay",
-            "type": "typeahead",
-            "length": "long",
-            "open_on_load": true
-        },
-        "biosample_ontology.classification": {
-            "title": "Biosample classification"
-        },
-        "biosample_ontology.term_name" : {
-            "title": "Biosample term name",
-            "type": "typeahead",
-            "length": "long",
-            "open_on_load": true
-        },
-        "biosample_ontology.organ_slims": {
-            "title": "Organ",
-            "type": "typeahead"
-        },
-        "biosample_ontology.cell_slims": {
-            "title": "Cell",
-            "type": "typeahead"
-        },
-        "award.project": {
-            "title": "Project"
-        },
-        "award.rfa": {
-            "title": "RFA"
-        },
-        "assembly": {
-            "title": "Genome assembly"
-        },
-        "date_released": {
-            "title": "Date released"
-        },
-        "date_submitted": {
-            "title": "Date submitted"
-        },
-        "lab.title": {
-            "title": "Lab"
-        },
-        "internal_status": {
-            "title": "Internal status"
-        }
-    },
-    "columns": {
-        "accession": {
-            "title": "Accession"
-        },
-        "assay_term_name": {
-            "title": "Assay name"
-        },
-        "assay_title": {
-            "title": "Assay title"
-        },
-        "target.label": {
-            "title": "Target of assay"
-        },
-        "target.genes.symbol": {
-            "title": "Target gene symbol"
-        },
-        "biosample_summary": {
-            "title": "Biosample summary"
-        },
-        "biosample_ontology.term_name": {
-            "title": "Biosample term name"
-        },
-        "description": {
-            "title": "Description"
-        },
-        "lab.title": {
-            "title": "Lab"
-        },
-        "award.project": {
-            "title": "Project"
-        },
-        "status": {
-            "title": "Status"
-        }
-    },
-    "boost_values": {
-        "@type": 1.0,
-        "accession": 1.0,
-        "alternate_accessions": 1.0,
-        "assay_term_name": 1.0,
-        "assay_term_id": 1.0,
-        "dbxrefs": 1.0,
-        "aliases": 1.0,
-        "award.title": 1.0,
-        "biosample_ontology.term_id": 1.0,
-        "biosample_ontology.term_name": 1.0,
-        "biosample_ontology.classification": 1.0,
-        "biosample_ontology.organ_slims": 1.0,
-        "biosample_ontology.cell_slims": 1.0,
-        "biosample_ontology.developmental_slims": 1.0,
-        "category_slims": 1.0,
-        "objective_slims": 1.0,
-        "type_slims": 1.0,
-        "assay_synonyms": 1.0,
-        "biosample_ontology.synonyms": 1.0,
-        "award.project": 1.0,
-        "award.name": 1.0,
-        "submitted_by.email": 1.0,
-        "submitted_by.first_name": 1.0,
-        "submitted_by.last_name": 1.0,
-        "lab.institute_name": 1.0,
-        "lab.institute_label": 1.0,
-        "lab.title": 1.0,
-        "possible_controls.accession": 1.0,
-        "possible_controls.alternate_accessions": 1.0,
-        "target.aliases": 1.0,
-        "target.genes.symbol": 1.0,
-        "target.label": 1.0,
-        "target.genes.dbxrefs": 1.0,
-        "target.organism.name": 1.0,
-        "target.organism.scientific_name": 1.0,
-        "references.title": 1.0,
-        "notes": 1.0,
-        "internal_tags": 1.0,
-        "doi": 1.0,
-        "biosample_summary": 12.0,
-        "description": 8.0
-    },
-    "changelog": "/profiles/changelogs/single_cell_experiment.md"
-}
