--- src/encoded/tests/data/inserts/high_throughput_single_cell_experiment.json
@@ -1,16 +1,0 @@
-[
-    {
-        "_test": "10X",
-        "accession": "ENCSR100TEX",
-        "assay_term_name": "single-cell RNA sequencing assay",
-        "award": "U54HG007004",
-        "biosample_ontology": "cell_line_EFO_0002067",
-        "description": "10X single-cell RNA sequencing assay experiment",
-        "lab": "thomas-gingeras",
-        "status": "released",
-        "date_submitted": "2015-01-20",
-        "date_released": "2016-01-01",
-        "submitted_by": "dignissim.euismod@amet.habitant",
-        "uuid": "11d378f3-aba4-128d-908b-d0d3f60d2e2e"
-    }
-]
