--- src/encoded/types/single_cell_experiment.py
@@ -1,379 +1,0 @@
-from pyramid.traversal import find_root
-from snovault import (
-    abstract_collection,
-    calculated_property,
-    collection,
-    load_schema,
-)
-from .base import (
-    paths_filtered_by_status
-)
-from .dataset import Dataset
-
-from .shared_calculated_properties import (
-    CalculatedAssaySynonyms,
-    CalculatedAssayTermID,
-    CalculatedVisualize,
-    CalculatedBiosampleSummary,
-    CalculatedReplicates,
-    CalculatedAssaySlims,
-    CalculatedAssayTitle,
-    CalculatedCategorySlims,
-    CalculatedTypeSlims,
-    CalculatedObjectiveSlims
-)
-
-from .assay_data import assay_terms
-
-@abstract_collection(
-    name='single-cell-experiments',
-    unique_key='accession',
-    properties={
-        'title': "Single cell experiments",
-        'description': 'Listing of all types of single cell experiments.',
-    })
-class SingleCellExperiment(
-    Dataset,
-    CalculatedAssaySynonyms,
-    CalculatedAssayTermID,
-    CalculatedVisualize,
-    CalculatedAssaySlims,
-    CalculatedAssayTitle,
-    CalculatedCategorySlims,
-    CalculatedTypeSlims,
-    CalculatedObjectiveSlims,
-    ):
-    item_type = 'single_cell_experiment'
-    base_types = ['SingleCellExperiment'] + Dataset.base_types
-    schema = load_schema('encoded:schemas/single_cell_experiment.json')
-    embedded = Dataset.embedded + [
-        'biosample_ontology',
-        'files.platform',
-        'files.analysis_step_version.analysis_step',
-        'files.analysis_step_version.analysis_step.pipelines',
-        'related_series',
-        'replicates.antibody',
-        'replicates.library',
-        'replicates.library.biosample.biosample_ontology',
-        'replicates.library.biosample.submitted_by',
-        'replicates.library.biosample.source',
-        'replicates.library.biosample.applied_modifications',
-        'replicates.library.biosample.applied_modifications.documents',
-        'replicates.library.biosample.organism',
-        'replicates.library.biosample.donor',
-        'replicates.library.biosample.donor.organism',
-        'replicates.library.biosample.part_of',
-        'replicates.library.biosample.part_of.donor',
-        'replicates.library.biosample.part_of.treatments',
-        'replicates.library.biosample.treatments',
-        'replicates.library.construction_platform',
-        'replicates.library.treatments',
-        'possible_controls',
-        'target.genes',
-        'target.organism',
-    ]
-
-
-@collection(
-    name='low-throughput-single-cell-experiments',
-    unique_key='accession',
-    properties={
-        'title': 'Low throughput single cell experiments',
-        'description': 'Listing of Low Throughput Single Cell Experiments',
-    })
-class LowThroughputSingleCellExperiment(
-    SingleCellExperiment
-    ):
-    item_type = 'low_throughput_single_cell_experiment'
-    schema = load_schema('encoded:schemas/low_throughput_single_cell_experiment.json')
-    embedded = [
-        'submitted_by',
-        'lab',
-        'award.pi.lab',
-        'biosample_ontology',
-        'related_series',
-        'possible_controls',
-        'target.genes',
-        'target.organism',
-    ]
-    audit_inherit = [
-    ]
-    set_status_up = [
-        'original_files',
-        'documents',
-        'target',
-    ]
-    set_status_down = [
-        'original_files',
-    ]
-    rev = Dataset.rev.copy()
-    rev.update({
-        'related_series': ('Series', 'related_datasets'),
-        'superseded_by': ('LowThroughputSingleCellExperiment', 'supersedes')
-    })
-
-    @calculated_property(schema={
-        "title": "Related series",
-        "type": "array",
-        "items": {
-            "type": ['string', 'object'],
-            "linkFrom": "Series.related_datasets",
-        },
-        "notSubmittable": True,
-    })
-    def related_series(self, request, related_series):
-        return paths_filtered_by_status(request, related_series)
-
-    @calculated_property(schema={
-            "title": "Superseded by",
-            "type": "array",
-            "items": {
-                "type": ['string', 'object'],
-                "linkFrom": "LowThroughputSingleCellExperiment.supersedes",
-            },
-            "notSubmittable": True,
-    })
-    def superseded_by(self, request, superseded_by):
-        return paths_filtered_by_status(request, superseded_by)
-
-
-@collection(
-    name='high-throughput-single-cell-experiments',
-    unique_key='accession',
-    properties={
-        'title': 'High throughput single cell experiments',
-        'description': 'Listing of High Throughput Single Cell Experiments',
-    })
-class HighThroughputSingleCellExperiment(
-    SingleCellExperiment,
-    CalculatedBiosampleSummary,
-    CalculatedReplicates,
-    ):
-    item_type = 'high_throughput_single_cell_experiment'
-    schema = load_schema('encoded:schemas/high_throughput_single_cell_experiment.json')
-    embedded = Dataset.embedded + [
-        'biosample_ontology',
-        'files.platform',
-        'files.analysis_step_version.analysis_step',
-        'files.analysis_step_version.analysis_step.pipelines',
-        'related_series',
-        'replicates.antibody',
-        'replicates.library',
-        'replicates.library.biosample.biosample_ontology',
-        'replicates.library.biosample.submitted_by',
-        'replicates.library.biosample.source',
-        'replicates.library.biosample.applied_modifications',
-        'replicates.library.biosample.applied_modifications.documents',
-        'replicates.library.biosample.organism',
-        'replicates.library.biosample.donor',
-        'replicates.library.biosample.donor.organism',
-        'replicates.library.biosample.part_of',
-        'replicates.library.biosample.part_of.donor',
-        'replicates.library.biosample.part_of.treatments',
-        'replicates.library.biosample.treatments',
-        'replicates.library.construction_platform',
-        'replicates.library.treatments',
-        'possible_controls',
-        'target.genes',
-        'target.organism',
-    ]
-    audit_inherit = [
-        'original_files',
-        'original_files.replicate',
-        'original_files.platform',
-        'target',
-        'files.analysis_step_version.analysis_step.pipelines',
-        'revoked_files',
-        'revoked_files.replicate',
-        'submitted_by',
-        'lab',
-        'award',
-        'documents',
-        'replicates.antibody.characterizations.biosample_ontology',
-        'replicates.antibody.characterizations',
-        'replicates.antibody.targets',
-        'replicates.library',
-        'replicates.library.documents',
-        'replicates.library.biosample',
-        'replicates.library.biosample.biosample_ontology',
-        'replicates.library.biosample.organism',
-        'replicates.library.biosample.treatments',
-        'replicates.library.biosample.applied_modifications',
-        'replicates.library.biosample.donor.organism',
-        'replicates.library.biosample.donor',
-        'replicates.library.biosample.treatments',
-        'replicates.library.biosample.originated_from',
-        'replicates.library.biosample.originated_from.biosample_ontology',
-        'replicates.library.biosample.part_of',
-        'replicates.library.biosample.part_of.biosample_ontology',
-        'replicates.library.biosample.pooled_from',
-        'replicates.library.biosample.pooled_from.biosample_ontology',
-        'replicates.library.spikeins_used',
-        'replicates.library.treatments',
-        'target.organism',
-    ]
-    set_status_up = [
-        'original_files',
-        'replicates',
-        'documents',
-        'target',
-    ]
-    set_status_down = [
-        'original_files',
-        'replicates',
-    ]
-    rev = Dataset.rev.copy()
-    rev.update({
-        'related_series': ('Series', 'related_datasets'),
-        'replicates': ('Replicate', 'experiment'),
-        'superseded_by': ('HighThroughputSingleCellExperiment', 'supersedes')
-    })
-
-    @calculated_property(schema={
-        "title": "Related series",
-        "type": "array",
-        "items": {
-            "type": ['string', 'object'],
-            "linkFrom": "Series.related_datasets",
-        },
-        "notSubmittable": True,
-    })
-    def related_series(self, request, related_series):
-        return paths_filtered_by_status(request, related_series)
-
-    @calculated_property(schema={
-            "title": "Superseded by",
-            "type": "array",
-            "items": {
-                "type": ['string', 'object'],
-                "linkFrom": "HighThroughputSingleCellExperiment.supersedes",
-            },
-            "notSubmittable": True,
-    })
-    def superseded_by(self, request, superseded_by):
-        return paths_filtered_by_status(request, superseded_by)
-
-
-@collection(
-    name='single-cell-units',
-    unique_key='accession',
-    properties={
-        'title': 'Single cell units',
-        'description': 'Listing of single cell units',
-    })
-class SingleCellUnit(
-    Dataset,
-    CalculatedBiosampleSummary,
-    CalculatedReplicates,
-    CalculatedAssaySynonyms,
-    CalculatedAssayTermID,
-    CalculatedVisualize,
-    CalculatedAssaySlims,
-    CalculatedAssayTitle,
-    CalculatedCategorySlims,
-    CalculatedTypeSlims,
-    CalculatedObjectiveSlims
-    ):
-    item_type = 'single_cell_unit'
-    schema = load_schema('encoded:schemas/single_cell_unit.json')
-    embedded = Dataset.embedded + [
-        'biosample_ontology',
-        'files.platform',
-        'files.analysis_step_version.analysis_step',
-        'files.analysis_step_version.analysis_step.pipelines',
-        'related_series',
-        'replicates.antibody',
-        'replicates.library',
-        'replicates.library.biosample.biosample_ontology',
-        'replicates.library.biosample.submitted_by',
-        'replicates.library.biosample.source',
-        'replicates.library.biosample.applied_modifications',
-        'replicates.library.biosample.applied_modifications.documents',
-        'replicates.library.biosample.organism',
-        'replicates.library.biosample.donor',
-        'replicates.library.biosample.donor.organism',
-        'replicates.library.biosample.part_of',
-        'replicates.library.biosample.part_of.donor',
-        'replicates.library.biosample.part_of.treatments',
-        'replicates.library.biosample.treatments',
-        'replicates.library.construction_platform',
-        'replicates.library.treatments',
-        'possible_controls',
-        'target.genes',
-        'target.organism',
-    ]
-    audit_inherit = [
-        'original_files',
-        'original_files.replicate',
-        'original_files.platform',
-        'target',
-        'files.analysis_step_version.analysis_step.pipelines',
-        'revoked_files',
-        'revoked_files.replicate',
-        'submitted_by',
-        'lab',
-        'award',
-        'documents',
-        'replicates.antibody.characterizations.biosample_ontology',
-        'replicates.antibody.characterizations',
-        'replicates.antibody.targets',
-        'replicates.library',
-        'replicates.library.documents',
-        'replicates.library.biosample',
-        'replicates.library.biosample.biosample_ontology',
-        'replicates.library.biosample.organism',
-        'replicates.library.biosample.treatments',
-        'replicates.library.biosample.applied_modifications',
-        'replicates.library.biosample.donor.organism',
-        'replicates.library.biosample.donor',
-        'replicates.library.biosample.treatments',
-        'replicates.library.biosample.originated_from',
-        'replicates.library.biosample.originated_from.biosample_ontology',
-        'replicates.library.biosample.part_of',
-        'replicates.library.biosample.part_of.biosample_ontology',
-        'replicates.library.biosample.pooled_from',
-        'replicates.library.biosample.pooled_from.biosample_ontology',
-        'replicates.library.spikeins_used',
-        'replicates.library.treatments',
-        'target.organism',
-    ]
-    set_status_up = [
-        'original_files',
-        'replicates',
-        'documents',
-        'target',
-    ]
-    set_status_down = [
-        'original_files',
-        'replicates',
-    ]
-    rev = Dataset.rev.copy()
-    rev.update({
-        'related_series': ('Series', 'related_datasets'),
-        'replicates': ('Replicate', 'experiment'),
-        'superseded_by': ('SingleCellUnit', 'supersedes')
-    })
-
-    @calculated_property(schema={
-        "title": "Related series",
-        "type": "array",
-        "items": {
-            "type": ['string', 'object'],
-            "linkFrom": "Series.related_datasets",
-        },
-        "notSubmittable": True,
-    })
-    def related_series(self, request, related_series):
-        return paths_filtered_by_status(request, related_series)
-
-    @calculated_property(schema={
-            "title": "Superseded by",
-            "type": "array",
-            "items": {
-                "type": ['string', 'object'],
-                "linkFrom": "SingleCellUnit.supersedes",
-            },
-            "notSubmittable": True,
-    })
-    def superseded_by(self, request, superseded_by):
-        return paths_filtered_by_status(request, superseded_by)
